Oops, fixing a mistake on the resendRequest function
This function was testing the presence of a lock instance PID or registred
name in the Refs list. If it was present, nothing should be done, otherwise a
new request message should be sent to guarantee that an ok message would be
received afterwards from a higher priority lock instance.