Opty split for multiple machines
Splitting the opty.erl module into two modules to make it possible
and the needed changes to make it possible to run. Fixing a possible
issue on the client code while it could be waiting for ever for an
answer from a Handler, that therefore is waiting for an answer from
an Entry that could be already dead, since they are managed by the
server module.